add module for core (creates user)

See merge request !86

+++ supervisr/puppet/builder.py
@@ -10,6 +10,8 @@ import sys
 import tarfile
 from tempfile import NamedTemporaryFile
 
+from django import conf
+from django.contrib.auth.models import Group
 from django.core.files import File
 from django.template import loader
 
@@ -20,16 +22,16 @@ from .utils import ForgeImporter
 
 LOGGER = logging.getLogger(__name__)
 
-
+# pylint: disable=too-many-instance-attributes
 class ReleaseBuilder(object):
     """
     Class to build PuppetModuleRelease's in Memory from files and templates
     """
 
     module = None
-    extension = '.djt'
     base_dir = None
     version = None
+    output_base = None
 
     _root_dir = ''
     _spooled_tgz_file = None
@@ -41,6 +43,7 @@ class ReleaseBuilder(object):
         self.module = module
         if self.module.source_path:
             self.base_dir = self.module.source_path
+        self.output_base = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'modules')
         # If version is None, just use the newest Release's ID + 1
         if version is None:
             releases = PuppetModuleRelease.objects.filter(module=module)
@@ -64,7 +67,10 @@ class ReleaseBuilder(object):
             'PUPPET': {
                 'module': self.module,
                 'version': self.version,
-            }})
+            },
+            'settings': conf.settings,
+            'puppet_systemgroup': Group.objects.get(name='Puppet Systemusers'),
+            })
         return context
 
     def to_tarinfo(self, template, ctx, rel_path):
@@ -74,16 +80,14 @@ class ReleaseBuilder(object):
         # First off render the template
         tmpl = loader.get_template(template)
         rendered = tmpl.render(ctx)
-        # Create the new path without the .djt
-        new_path = rel_path.replace(self.extension, '')
         # If it's a json file now, check if it's valid
-        if new_path.endswith('.json'):
+        if rel_path.endswith('.json'):
             self.validate_json(rendered)
-            LOGGER.info('Successfully validated %s', new_path)
+            LOGGER.info('Successfully validated %s', rel_path)
         # Convert it to bytes, create a TarInfo object and add it to the main archive
         byteio = io.BytesIO(rendered.encode('utf-8'))
         byteio.seek(0, io.SEEK_END)
-        tar_info = tarfile.TarInfo(name=new_path)
+        tar_info = tarfile.TarInfo(name=rel_path)
         tar_info.size = byteio.tell()
         byteio.seek(0, io.SEEK_SET)
         self._tgz_file.addfile(tar_info, fileobj=byteio)
@@ -132,7 +136,7 @@ class ReleaseBuilder(object):
             return matches
 
     @time
-    def build(self, context=None, db_add=True):
+    def build(self, context=None, db_add=True, force_rebuild=False):
         """
         Copy non-templates into tar, render templates into tar and import into django
         """
@@ -141,12 +145,12 @@ class ReleaseBuilder(object):
             context = {}
         _context = self.make_context(context)
         for file in files:
-            # Render template if matches extension
+            # Render template
             arc_path = file.replace('\\', '/').replace(self.base_dir, self._root_dir + '/')
-            if arc_path.endswith(self.extension):
-                self.to_tarinfo(file, _context, arc_path)
-            else:
+            if os.path.isdir(file):
                 self._tgz_file.add(file, arcname=arc_path, recursive=False)
+            else:
+                self.to_tarinfo(file, _context, arc_path)
             LOGGER.info('Added %s', arc_path)
 
         # Flush to file buffer
@@ -154,8 +158,8 @@ class ReleaseBuilder(object):
         # Gzip it so we actually have a tgz
         gzipped = gzip.compress(self._spooled_tgz_file.getbuffer())
         # Write to file and add to db
-        module_dir = 'puppet/modules/%s/%s/' \
-                     % (self.module.owner.username, self.module.name)
+        module_dir = '%s/%s/%s/' \
+                     % (self.output_base, self.module.owner.username, self.module.name)
         prefix = '%s-%s_version_%s_' % (self.module.owner.username, self.module.name, self.version)
         if not os.path.exists(module_dir):
             os.makedirs(module_dir)
@@ -163,12 +167,13 @@ class ReleaseBuilder(object):
             with NamedTemporaryFile(dir=module_dir, suffix='.tgz', prefix=prefix) as temp_file:
                 temp_file.write(gzipped)
                 temp_file.seek(0, io.SEEK_SET)
+                LOGGER.info("Target filename: %s", temp_file.name)
                 # Create the module in the db and write it to disk
                 self._release = PuppetModuleRelease.objects.create(
                     module=self.module,
                     version=self.version,
                     release=File(temp_file))
-        else:
+        elif force_rebuild is True:
             with open(prefix+'.tgz', mode='w+b') as file:
                 file.write(gzipped)
                 LOGGER.info("Wrote module to %s", prefix+'.tgz')
